refactor(organisation): create src folder
